LayoutTests/imported/w3c/web-platform-tests/content-security-policy/unsafe-hashes/javascript_src_allowed-window_location-expected.txt


FAIL Test that the javascript: src is allowed to run assert_equals: expected "pass" but got "fail"
